Add io_get_scan_result, remove large binaries, add notes to compile
Replaced using the deployed jars from BDD6502 with instructions
to build locally. This means more work for people using this, but
gives full source access for writing new code, and feature completion
in IntelliJ.

Should further improve when BDD6502 is released as a repository, then
all the maven instructions in here can be removed.


Former-commit-id: 3b666300a262780fadc71dc21519143321770cc3 [formerly e728476219405aed7bc4a8bee28e7f12d0f6e8c9]
Former-commit-id: 88ec54fe903a6c233ccc8f573e2c84540a692249